When all entries have finished loading, update the total unread count in the favicon.
In 7adc056d72fb4b3ef331aa79dd29ceba5fe22861 readSvc was changed so that unread counts in feeds were updated when entries pagination was finished. However the total unread count in the favicon was not updated as well, which could cause a discrepancy between the count reported in the sidebar (and window title) and the count reported in the favicon.

To fix this, the change in the unread count in readSvc is not done accessing the unread_count attribute in feed objects directly, but rather invoking a method in changeUnreadCountSvc. This method updates the unread_count attribute in feed objects, and it also takes care of invoking favicoSvc to update the favicon count.

Generally, changes in the unread count of feeds should be done through changeUnreadCountSvc, so that any extra actions (like updating the favicon) are handled transparently.