WIP: show product tours.
Currently the sample tour from the Hopscotch docs has been implemented.

Not yet implemented: show the tour only if the user has not yet seen it; this means adding database columns for this information and AJAX calls to retrieve it client-side. Update these new columns when the user completes the tour or closes it, again with AJAX calls.

The images for step numbers, tour closing icon etc have been moved to the public directory because the Hopscotch library looks for them in the /img path, which does not mix well with the asset pipeline's /asset path.