Tour i18n strings are retrieved from the server via AJAX, instead of hardcoded in the javascript client code.
This way Rails i18n mechanisms can be used to display tours in the user's locale. Maintenance will be easier using a single i18n mechanism for the whole app. Also the javascript client code is the same for all users, regardless of their chosen locale, which helps with caching.

The tour strings are currently hardcoded in the Rails view that generates the JSON response. Next few commits will switch to i18n-enabled strings.