Added update method to UserConfigsController.
Currently it only accepts a "true" or "false" string for show_main_tour. It updates the show_main_tour attribute of the current user accordingly.

This is intended so that the flag is updated when the user completes or closes the main tour, so that it is not shown again.