Removed "Back" button from all tours.
It was too much trouble. Because of the dynamic nature of the app pages, elements could move or disappear which could cause tour bubbles to be misplaced or missing when going back. It's simpler to just remove the Back button. Tours can be reset from the profile page if the user wants to go through them again, anyway.