Simplified target of a step in the main app tour. This required reworking the show/hide read entries button in the navbar.
This shouldn't be perceptible for the user, but now the tour step targets an element that is always visible (the <li> that wraps both buttons, only one of which is visible at any given time). This fixes some potential placement issues with the tour bubble placement.

An aditional option (container: "body") has been added via data- attributes for the boostrap tooltips of both buttons. This is necessary because otherwise bootstrap uses a too small width for tooltips, which would look ugly.