In the angularjs findSvc service, when searching for the "all" or "none" folders, a result is returned even if the user currently has no folders defined.
This is necessary to fix a bug: when the user had no folders, if he clicked on the "all subscriptions" link, no entries were loaded. The reason is that findSvc returned null instead of the object for the "all" folder. The condition to only return results if there is a list of folders in the root scope is valid when then service is asked for a user-created folder, but not when asking for the "all" or "none" folders, which exist resgardless of whether the user has created folders or not.

Also a regression test for this bug has been implemented.