When subscribing to a feed, the unread count in favicon and window title are updated after the feed has been loaded from the server.
This was done in the wrong place. Before this commit, the favicon/title update happened just after the client was notified that the subscribe job had finished successfully. At this moment it is likely that the feed has not been loaded from the server yet, which means that its unread entries would not be considereed for the total unread count. This is not a big bug, because the count will be updated in the next scheduled client update (1 minute later at the most), but annoying nevertheless.