Height of the feed title link (above the entries list) is now responsive.
For tablet or smartphone-sized screens, it has a bigger height, to match the increased height of other elements (sidebar links, entries) in small screens. This makes it easier to use in small touchscreens.

Also vertical height (achieved with responsive top and bottom paddings) is calculated so that the hr line under the feed title is always aligned with the hr line under the Start link in the sidebar, across all viewport widths.