The responsive vertical size of feed titles implemented in 1c2229c085291ecd2223e27fbf707d0d1e082b27 is also applied now to folder titles and to the "All subscriptions" text, when appropiate.
When the user has selected "all subscriptions" or a whole folder for reading, the same styling should be applied to the title above the entries list as if the user had selected a single feed for reading. I forgot to implement this in 1c2229c085291ecd2223e27fbf707d0d1e082b27.