Alerts are shown at the bottom of the page, instead of being shown at the top.
This way they are still visible but they never overlap the navbar, which is crucial for the usage of the app.

This required some reworking of the cookies warning, which is now shown nested into the same div as other alerts. This is necessary so that any other alerts do not overlap the cookies warning.

The only alert still displayed at the top of the page is the noscript warning.