Removed OPML import partial from the pages layout.
Currently there is no way to show this popup while on a high_voltage page. It is a remnant from when the OPML import could be started from the user menu; now that it can be invoked either from the start page or from the edit registration view, it is dead markup that gets downloaded but is never shown in static pages.