Configure Rails in the development environment to eager-load code in the configured paths (which by default includes everything under /app) on startup.
In particular, this means that (among other things) Sidekiq workers are loaded on startup. This is necessary for the workers to appear in the Sidetiq tab of Sidekiq's web UI. The default Rails 4 configuration disables eager loading in development, which means that the mentioned Sidetiq tab does not display anything.

The price to pay is slower Rails startup in development. Oh well.

For more about this see:

https://github.com/tobiassvn/sidetiq/wiki/Known-Issues

and the lengthy discussion about the issue in:

https://github.com/tobiassvn/sidetiq/issues/12