WIP: adapt the "fix missing feed update schedules" feature to Sidekiq.
As part of the migration, some method names have been changed to better reflect the new behavior. It should be transparent for the end user, but now feed updates are no longer made recurrent by a gem. Instead, the next update of a feed is scheduled, and each time an update runs it is responsible to schedule the next feed update at the correct time.