Update Redis config to redis-server version 2.8.17
The production server was using a very old redis-server 2.2. Sidetiq uses some redis commands that are only supported in version 2.6+

It really made no sense using such an old redis-server version, so I've updated redis version in all machines (dev, prod) to 2.8.17 (latest stable). As usual config options and defaults change from one redis version to the next, so I've updated the config file commited to git so that it's aligned with the default config for this version.

Of course in production a few options are different, because this file is not deployed automatically by capistrano, it is linked from the shared folder (where it must be changed by hand).