Minor change in feed title link width, to fix a positioning issue with its tooltip.
If we make the link have a fixed width of 100%, the tooltip is positioned in the middle of that width, which means that it's away from the text if the feed title is short.

The fix is to make it a maximum of 100% width, but if the text is shorter, the width is smaller. This way the tooltip is correctly positioned.