Upgraded favico.js 0.3.4 -> 0.3.5. Also favico.js is added to this app's asset pipeline, instead of being served from the CDNJS cloudflare CDN.
I've been having some SSL issues with CDNJS, I'm not sure it's too reliable. favico.js is not a very popular library and it's not hosted in the major public CDNs, also its developer does not recommend a CDN. It is present in CDNJS but it doesn't get much love (version 0.3.5 was released three months ago but it's not yet available in CDNJS).

It's a rather small library and I've decided it's better to add it to the asset pipeline so that it's served from our servers instead of a CDN.