Added task to remotely restart Puma via Capistrano (not just issue a phased restart, but a full restart).
Remember: Our Capistrano script issues a phased restart after a deployment. A full restart is necessary after any changes to the db schema, or any changes which would make a phased restart dangerous (any changes in which outdated code still bein executed in a not yet restarted worker would be dangerous).

The idea is that after a deployment including changes incompatible with a phased restart, a "cap puma:restart" will be issued to ensure a full restart.

Note: with a phased restart users will see no disruption, but even with a full restart no connections are dropped, at most users will experience some delay while the app is bein reloaded.

For more about when a phased restart is dangerous and a full restart is recommended, see:

https://github.com/puma/puma#normal-vs-hot-vs-phased-restart