Upgraded Rails 4.1.6 -> 4.1.7
This fixes a vulnerability in Rails and Sprockets when serving static assets.

It's not a huge deal, and it probably doesn't affect this app because static assets in production are served by nginx, not by Rails. But it's always better to be up to date with security releases.

For more about the vulnerabilities fixed see:

http://weblog.rubyonrails.org/2014/10/30/Rails_3_2_20_4_0_11_4_1_7_and_4_2_0_beta3_have_been_released/