When a user queries for the unread entries count for a feed, if he's not subscribed to the feed, return zero instead of rising an error.
This should fix the following issue: when unsubscribing from a feed, the actual unsubscribing happens asynchronously in a sidekiq worker. A possible sequence of events is:

- the user requests unsubscribing from feed F. UnsubscribeUserWorker is enqueued in Sidekiq
- the browser decides to refresh feeds and unread counts by making an Ajax request (this happens every minute as long as the read view is open, and exactly when it happens is out of the user's control).
- a FeedController instance retrieves starts composing the JSON response. As the first step it retrieves from the DB all subscribed feeds; in this moment the user is still subscribed to feed F, so it is present in the set of feeds.
- the UnsubscribeUserWorker instance runs, and its perform method unsubscribes the user. Note that this is asynchronous to the FeedController instance running above.
- the FeedController instance continues running and starts iterating over the subscribed feeds (which, rememember, include feed F) to compose the JSON response. When it gets to processing feed F, the user is no longer subscribed to it.

The old behavior was to raise an error when the program reaches this point. This meant that instead of returning JSON, an HTTP error was returned, which made an error alert be displayed to the user. This was surprising for the user, to say the least.

The new behavior is to return zero. This means that no error is displayed to the user.

Currently the already unsusbcribed feed will appear in the sidebar for yet another minute, until the next automatic update. This should be fixed in the future, so that these feeds are not included in the returned JSON.

Also, the case in which the feed has not only been unsubscribed but also deleted from the database (because there were no more users subscribed) is not properly controlled. An error is likely still raised in this case; this should be fixed in the future.