When returning to the user JSON with the subscribed feeds, if a feed becomes unsubscribed during the process (e.g. because of a sidekiq worker running asynchronously), ignore it (it won't appear in the returned JSON) and continue with the next feed, insted of completely failing the FeedsController#index action and returning an HTTP error.
This should help fix the issue described in commit 82a00f4f6fc49078cd38dbab8a8df1118135aef1: when unsubscribing from a feed, if an automated refresh of feeds in the client happens just at the same time that the UnsubscribeUserWorker instance was running, it was possible (depending on timing) that the user was shown an error alert ("error retrieving feeds information" or some such).

A desirable side effect of the changes here is that views involved in FeedsController#index and FeedsController#show no longer access the database. Any database access is performed at the controller level, and the views are kept as simple as possible; they just receive simple values and compose the JSON response with them.

Also a minor change has been made that should improve performance: when retrieving feeds in a specific folder (e.g. when the user clicks on the "all subscriptions" link under a folder), the app no longer queries the database to find out in which folder is each returned feed. In this case we already know that all feeds are in that specific folder, we can just use its ID without querying the DB or performing any other logic.