Require Sidekiq >= 3.2.5
Version 3.2.4 and below have issues when shutting down. There's also mention in the bugreport of a memory leak, but I haven't been able to confirm it in this app.

The fix is to add a >=3.2.5 condition in the Gemfile. That is the first version that fixes this issue, or at least works around it. It seems to be some problem with recent Celluloid versions, or a race condition that only manifests in recent Celluloid versions. Version 3.2.5 of Sidekiq fixes version 0.15.2 of Celluloid as a dependency, which seems to make the problem go away for now.

I don't directly include Celluloid, so I'm not too worried about staying with older versions as long as Sidekiq works.

For more about this bug:

https://github.com/celluloid/celluloid/issues/457