Removed code that decided whether to apply a CSS class that added padding-top to the body to leave enough space for the navbar.
It is not necessary to check if the current user is authenticated to apply the class. There are views which only an authenticated user can reach, and it's safe to always use this class (the navbar is always visible when a user is authenticated). Conversely there are views which can only be reached by unauthenticated users; it's safe never using the class in those, because the navbar is never visible for unauthenticated users.