Added rake task and capistrano task for clearing the Rails cache.
The rake task can be invoked with "rake rails_cache:clear" to clear the cache in the local machine.

The capistrano task can be invoked with "cap production redis_cache:clear" to remotely clear the cache in the production environment. This actually invokes remotely the above rake task.

The idea is to invoke this if the server starts serving stale content, because of changes to views that are not picked by the caching system or any other reason.