Refined caching of the user menu in the navbar.
Besides caching different versions of the fragment for each locale, different versions are cached for the read page (without a "read feeds" menu option) and for other pages (with a "read feeds" menu option).

This should restore the old behavior: the "read feeds" option only appears in the menu if the user is not currently in the read page. It makes no sense having an option that just loads the page the user is already at.