Added caching and refined authentication token in forms that are part of the edit registration view.
The invite friend form is an Ajax form, so no authentication_token field is inserted. The destroy profile form is a regular form, so the authentication token field is present with an empty string for value; the jquery_ujs library will update this field with the correct CSRF token extracted from the page header.