WIP: when certain changes are made, update the updated_at attribute of subscriptions.
Specifically:

- when a feed title changes
- when a feed url changes
- when a feed is moved into a folder or out of a folder

This way when any of those changes is done, the updated_at attribute of associated FeedSubscription instances is updated to the current date, invalidating the HTTP cache in FeedsController#show. This means that after any of those changes, when a user next requests JSON describing the feed, a new JSON is returned with a 200 status, instead of returning a 304.