Added subscriptions_updated_at attribute to Users model.
The idea is that this attribute will be updated every time there is a change in users subscriptions; that is, a change that would change the JSON returned by FeedsController#index. This way, HTTP caching can be implemented in FeedsController#index by using this new attribute as last_modified value.

Implemented: added attribute. It takes a default value of the current date/time when a User is created.
Not implemented: actually update the attribute when subscriptions are updated.