Removed NOT NULL constraint in subscriptions_updated_at attribute of User model.
It seems that the invite! method added by devise_invitable to the User model saves users in the db without going through the before_validate callback. This means that every time an invitation is sent, a User is saved in the db without passing a value to this attribute; this raises an ActiveRecord error if the attribute has a NOT NULL constraint.

The fix is to remove the constraint so that devise_invitable can save invited users with the subscriptions_updated_at attribute set to null. This is not a huge issue because after that users should pass through the before_validate callback, which sets the attribute to the current date/time if it has a null value.