When caching the footer_js fragment, do not include in the cache the line that includes the application script.
If we cache that line, the resulting HTML will include a specific version of the application javascript (built by the asset pipeline), and future changes to javascript won't be served to users because the HTML will keep pointing to an old version of the js. Worse, there could be a moment in which the HTML pointed to an application.js version that has already been deleted from the server.

The fix is to leave that line outside the cache, and therefore dynamic. Another possible fix would have been to leave it cached and clear the cache by hand every time there is a change in the application javascript, but that's probably worse.

Changes to the rest of the libraries (e.g. an update to the Angularjs version) are not a problem because Rails includes the fragment hash as part of the key in the cache. This means that if we ever update e.g. Angularjs, the fragment hash will change, a new key will be generated and the old cached fragment will become stale.