When caching the header fragment, do not cache the line that imports the application CSS file (built by the asset pipeline).
The reasons are the same as in commit ca187bba335381f6484440639b6a5e624ee2731c: if we cache this line, the version of the CSS served to users would be fixed and no changes to the CSS would ever be served to users. We need this line to be dynamic.