Fix bug: update User instance from RefreshFeedJobState AR callback only if user exists.
When RefreshFeedWorker detects that the passed user does not exist, it deletes the RefreshFeedJobState from the database before exiting. This invokes the after_destroy callback of the RefreshFeedJobState model, which tries to update the refresh_feed_jobs_updated_at attribute of the associated User.

The bug was that, in this particular case, the User instance does not exist (the user_id attribute is set to null), which raised an error (because the update method was invoked on a nil object). The fix is, in the callback, to check that the User instance exists before attempting to update it.