Blacklist unnecessary external URLs during testing, to speed up acceptance tests.
During testing fonts, facebook and twitter js, etc are not necessary and tests can be sped up a bit by blacklisting those URLs. Blacklisting, in the context of capybara-webkit, means that when the browser downloads the .js, .css or whatever file from that URL, an HTTP 200 is returned with an empty body, without actually ever hitting the network.

Inspired by this blog post:

http://robots.thoughtbot.com/speed-up-javascript-capybara-specs-by-blacklisting-urls