For forms that are submitted via Ajax through the ng-submit angularjs directive, pass the submit event as argument to the invoked function, and use it to prevent normal form submit at the end of the function.
This is necessary because of recent changes in how ng-submit works. Previously it stopped normal submit (via POST) of all forms in this app that include the directive, which is the desired behavior. However now it doesn't prevent normal submit if the form has an "action" attribute, even if its value is empty.

This is a big problem because simple_form adds an action="" attribute to generated forms. I haven't found a way to prevent this attribute from being generated, it's probably not supported by simple_form. This means that these forms are broken (the desired submit via Ajax works, but then a normal HTTP POST is sent which fails, because the route it is trying to POST to is not even enabled in router.rb).

The fix is to pass the submit event as argument to all functions that perform Ajax form submitting. The event is used just before returning to prevent the default form submit.

This is at best ugly, but necessary right now. I'm not happy with angularjs, I'm less happy with angular changing its behavior with no warning, but I'm even less happy with simple_form. It would seem that having full control over all form attributes should be a given, but it's not supported (or at least not documented). I'm probably going to get rid of simple_form sooner than later, it's an extremely convoluted solution to what should be a simple problem.