Use Poltergeist (based on the headless webkit-based PhantomJS browser) for testing, instead of capybara-webkit.
It is supposed to be faster, we'll see.

Phantomjs does not need a virtual framebuffer, so the lines that started xvfb in Travis CI have been removed.

Poltergeist currently does not support url blacklisting, which we added back in commit 7ba7af3f7af11b90a754769ac36824f53817b2cb, so the lines that configured url blacklisting have been commented. There is a PR waiting merge in the Poltergeist github repo, so these lines will probably be uncommented again in the future.

For the moment Poltergeist is configured not to raise Ruby errors when a javascript error happens, to ease transition from capybara-webkit to poltergeist. This may change in the future.