Use rack-cache middleware (through rest-client-components gem) to transparently cache HTTP client-side when fetching feeds.
This replaces the mechanism baked into feedbunch, based on the "etag" and "last_modified" attributes of the Feed model. Those attributes and the code that references them will be soon removed.

The solution based on rack-cache is preferable because:

-  better standards support (the solution baked into feedbunch did not support the cache-control header or other complex scenarios)
- less complexity (HTTP 304 responses no longer have to be treated as a special case; they are served to FeedClient as HTTP 200 responses, but with a body retrieved from the local cache instead of sent over the network)
- less code maintenance, shorter tests (I will trust rack-cache just works, so tests related to HTTP caching can be removed)
- less buggy, probably. Less face it, a rack middleware widely used has probably seen better debugging than my single-dev code

For the moment file storage is used both for metadata and entities, in the (Rails.root)/rack_cache/entitystore and (Rails.root)/rack_cache/metastore folders. In the future I may want to store metadata in memory, probably in a Redis instance (rack-cache supports memcached natively, but I prefer using another Redis instance instead of introducing memcached in the mix; there are gems that support using Redis as a backend for rack-cache).