Translucent grey blocks in the landing page.
The nice background is now partially visible behind the grey divs. It looks nice.

The acknowledgements block has less opacity than the main action block. I hope it subtly gives more weight to the really important part of the page, and that actionable UI elements (signup and signin links, mostly) are more important to the eye.