Remove folder_id argument from SubscribeUserWorker#perform
It is not necessary to pass the folder id anymore. Since recent changes, SubscribeUserWorker is enqueued only when the user manually requests a new subscription; this means that a nil was always being passed in this argument. Better to get rid of it.

This makes a private method so simple that it's been removed and its code incorporated to the main perform method.