When enqueing a job to import an OPML file, any failures data related to a previous import for the same user is destroyed.
This way it is guaranteed that any OpmlImportFailure instances in the db are related to import failures that happened during the last import process for a user. No information is retained for other import processes than the last one for any given user.

To achieve this, when enqueing an ImportSubscriptionsWorker, the old OpmlImportJobState for the user is destroyed; thanks to ActiveRecord callbacks, this triggers the destruction of any associated OpmlImportFailure instances. Immediately afterwards a new OpmlImportJobState is created with state RUNNING. This replaces the old behavior in which the OpmlImportJobState instance was updated, instead of destroyed and recreated.