Assign ImportSubscriptionsWorker to a lower priority Sidekiq queue.
OPML importing has higher priority than maintenance tasks, but lower than truly interactive tasks like subscribing to a feed. In as much as possible, I'd like users not to have interactive tasks delayed because other users are importing OPMLs. The user is warned that an OPML import is a slow process and he should not be surprised.

Of course a delay in interactive tasks is still possible if all sidekiq threads are used up by instances of ImportSubscriptionsWorker, but this is not fixable without dedicating threads to the interactive queue exclusively, which would add a complexity I want to avoid.