When saving import failures, strip the URL before saving.
This ensures that only the significant part of the URL, without leading or trailing whitespace, line breaks etc, is saved in the database. This in turn means that the URLs can be inserted into a JSON response without modification (e.g. a trailing \n in a URL would mean an invalid JSON response otherwise).