Configure Sidekiq as backend for ActiveJob.
This means that emails sent with deliver_later are sent in a sidekiq job. This requires creating a new queue "mailers", because ActiveMailer enqueues jobs in a queue with that name. As far as I can see ActiveMailer does not allow configuring a different queue.

Sidekiq jobs have not been migrated to the ActiveJob framework because some Sidekiq features are not yet supported there. Most important, Sidetiq (for recurring jobs) is not supported, which is a big problem. If ActiveJob ever supports recurring jobs, it may be worth looking at it again.