Include responders gem in bundle.
Since Rails 4.2 this gem is necessary to get support for class-level respond_to invocations. We use them extensively to restrict supported formats in controllers (mostly json for API controllers, html for others).

The gem was already in the bundle because devise pulls it, but it's better to require it explicitly.