Use respond_with in controllers as much as possible, instead of render. If an unsupported format (json, html...) is requested, return an HTTP 406.
Using respond_with is useful because it responds only to formats specified in a previous respond_to call at the controller class level (all controllers already had respond_to calls specifying desired formats, but they weren't actually doing anything). If an unsupported format is requested, an ActionController::UnknownFormat error is raised, which ApplicationController converts to a 406 response.

On the other hand, using render in controllers wasn't ideal because the render method does not respect the formats passed to respond_to. This means that when an unsupported format was requested, render tried to find a template for that format and, when it wasn't found, an error was raised that resulted in an HTTP 500 response. This was not ideal; it's better to return a 406 to let the client know that the problem is in the unsupported format, instead of a generic 500 error.

Also respond_with is less verbose than render.

The only drawback would be if we used template names different from their controller actions; this is not supported by respond_with as far as I see. However this was not the case in this app, and in any case would be easily fixed.

Partially inspired by this blog post:

  http://www.justinweiss.com/blog/2014/11/03/respond-to-without-all-the-pain/